split classes into their own files
I hate when files contain multiple class def
